pyproject.toml
[project]
name = "alpha-trader"
version = "0.1.0"
description = "Autonomous stock trading agent with professional risk management"
requires-python = ">=3.12"
dependencies = [
    "alpaca-py>=0.33.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "jinja2>=3.1.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "apscheduler>=4.0.0a1",
    "openai>=1.50.0",
    "pandas>=2.2.0",
    "pandas-ta>=0.3.14b1",
    "yfinance>=0.2.40",
    "python-telegram-bot>=21.0",
    "pydantic-settings>=2.7.0",
    "httpx>=0.28.0",
    "sse-starlette>=2.0.0",
    "python-multipart>=0.0.18",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 100
